requirements.txt
django==4.1.6
djangorestframework==3.14.0
pandas==1.1.3